Move CreationService tasks into a background job
